11051 FIX crash in notifications if "Exclude Service Groups (regex)" is used
If a notification rule used the option "Exclude Service Groups (regex)" and
no service group was defined for a service the notification rule processing
crashed. In the var/log/notify.log a traceback that contains the line
"TypeError: 'NoneType' object has no attribute '__getitem__'" was shown.
This is fixed.

CMK-5015

Change-Id: Ie6e8b5809ae55d196d47acac74fd9384002b52be