Introduce page state
The area on the top right of the page is now named "page state". According
to the new UX concept it contains an icon which a) indicates the area of
the Web GUI one is working on and b) shows some state of the page next to
some descriptive text and useful links.

Besides the general mechanism this also adds the page state for the
setup pages.

Change-Id: Id234c96725422b53f0aba70ef45dc9f29cad868a