Page state: Add reload indicator
The default page state are now also has an icon. It visualizes the
remaining time of the active reload timer.

To make this possible we had to change the reload scheduler to work with
multiple ticks on the way to the actual page reload. Each tick updates
the state of the reload indicator.

Change-Id: I93185dafcb4773c05b2f57227baaa0d102ad4fc1