Prepare to separate avoption form from request processing
The previous code mixed up rendering of the form and gathering the
submitted form from the HTTP request. This commit prepares for the
separation. In the next step the request processing is moved directly
into get_availability_options_from_request while the form rendering
will stay in _show_availability_options.

Change-Id: I3cfab14126fcd049d4184b730fb98f144ff0052c