11261 FIX Fix performance regression caused by too many live status queries between EC and core
The version 1.6.0p14 introduced an issue affecting the Event Console and it's
Livestatus communication with the local monitoring core.

Instead of querying static configuration related information, which is needed
by the Event Console only once per core restart, these information were not
cached as intended. This resulted in these queries being made over and over
again.

The query was executed in the following situations:

<ul>
<li>Multiple times when querying the "eventconsolestatus" table (Once for each host known by the Event Console)</li>
<li>Once for each created event</li>
</ul>

Change-Id: Ibcfad43bfdce4745e9215bd3417b622a74a845a1