Cleanup over generalized edit_valuespec
This function was only used by a single call site: The pagetype edit page.
Since this was only used once and had several unused features, the
needed subset has now been integrated into the pagetype edit page.

This function is far away from being ideal, but for now:

a) The processing is easier to understand
b) Similar to all other value spec based edit pages

So if you know that scheme, the processing is easier to understand than
before.

Along the way solving the original problem: Move the save button to the
page menu logic

Change-Id: I83d61b91c5fb0f76fcfd8c9fbc0d150d6f5d4964