11262 Activate changes: Asking for activation comment can now be configured
The comment that can be added before activating the pending changes is now
configurable using the global setting "Ask for a comment for activation of changes".

You can either choose to enforce your users to provide a comment, ask for an
optional comment or completely skip the question for a comment.

The default has been set to: "Do not ask for a comment".

Change-Id: I29788517421f99a56e6878d63aade22f618cdd0b