Simplify return type to `DataSource._get_raw_data()`
The highly optimized code in abcdatasource._run() prevents us from
moving FileCache handling to the fetchers.

In the new design, the raw data is provided by the fetchers.  Caching is
to be handled in the fetchers as well.  The checkers on the receiver end
must treat cached data and data from I/O in the same manner, that is,
parse them.  Neither end of the data chain has room for a boolean flag.

In this patch, we remove the boolean value returned by `_get_raw_data()`
to indicate whether the raw data comes from cache or I/O.  Instead,
we check whether the persisted sections are the same as the cached ones.
This should be functionaly equivalent.

CMK-5174

Change-Id: I104820cfa79223c6f74c34f724a7b8cfca321d56