Make WATO mode breadcrumbs contain their topics
For the top level modes we need to prepend the topic of the mode.  The
mode is sadly not available directly in WatoMode. Instead it is
configured in the MainModule class that is related to the WatoMode.
There is no 1:1 connection between WatoMode / MainModule classes.  For
the moment we lookup the main_module_registry to find the topics for as
many modes as possible.

Once all non top level modes have a parent_mode() method, we know which
modes are top level modes. Then we could move all attributes from the
MainModules to the WatoModes and create the WATO menu items directly out
of the WatoModes.

Change-Id: I1453a07f3f408ce6abc7a918a1a50c6353906252