Add conductor base class
Until now all quicksearch plugins are based on livestatus queries
and may be aggregated by the LivestatusSearchConductor. Once we
add the first non livestatus filter, we need to handle it in a
different way.

This commit prepares for introducing another type of plugin. This
will then be managed by a search conductor without further
aggregation.

Change-Id: Ie78bb290c454f80063bac142b63ee21a9ac5d694