Split check_plugins to 3 separate modes
Previously the check_plugins mode was implementing the
top level check catalog page, the topic lists and the
check plugin search. Nearly every function had conditions
to render either the one or the other logic.

To simplify things and make the breadcrumb easier to
implement, this commit splits up all of them into a
dedicated mode.

Change-Id: Ie8331e85592405190e6e91798a35f522e410cd22