Persist suggested actions globally
The state of the suggested actions that can be toggled e.g. in the
"Add host" page are now persisted across page reloads. Furthermore,
the state is now saved globally for all pages. By default the
suggested actions are now active i.e. the actions are shown as
large buttons.

CMK-5167

Change-Id: Ied5d1b7d667b2d231f05ef41ce1bb091e9de747c