automations/check_mk gets raw data from fetchers
This clarifies the intent.

This is also the first step toward removing the `run_raw()` method
from the data source.

The corresponding test was moved from `test_cmk_caching.py` to
`test_automations_check_mk.py` and completed such that the
return value of the call is actually checked.

CMK-5243

Change-Id: Id4a612b348ac8e2d45d5006b95e33bcf2dc6d5ef