automations/check_mk: merge loops
The code was iterating twice over the same collection in a few lines.
This is less efficient, harder to understand, and harder to
modify/maintain.  This patch merges the loops.

This is further required to remove the `run_raw()` call.  Removing the
`run_raw()` call is necessary to complete the split between fetchers and
checkers:  Indeed, there should not be a checkers method that produces
the raw data.

CMK-5243

Change-Id: I4041f17c73d00e3a66ed173b53d85c29586da56f