modes/check_mk: get raw data from fetcher
Same transformation as automations/check_mk

Now, the responsibilities of the fetchers and checkers are clearly
separated.  We can remove the last cruft out of checkers in a
next commit.

This further aids understanding the API.  Indeed, we could, for example,
rename the Configurator -> Source (or DataSource) and let

```
class Source(Generic[RawData]):
    def fetch() -> RawData:
        with self.make_fetcher() as fetcher:
	    return fetcher.fetch()

    def summarize(raw_data: RawData) -> ServiceCheckResult:
        checker = self.make_checker()
        checker.check(raw_data)
        return checker.get_summary_result()

```

which would be clearer and more concise while retaining a clear
separation of concerns.

CMK-5243

Change-Id: Ic7278fffba16b26d009293eb44eb94b889cdd923