Clean up DataSource._run()
The method was using the boolean-parameter antipattern to let the method
return either raw data or host sections.  Now, checkers should never
return raw data as this is now the responsibility of the fetchers.

Furthermore, the AgentDataSource.run_raw() does not have callers
anymore so that it is safe to remove the faulty logic.

CMK-5423

Change-Id: I4d4e5dc8e4e57ba02db7681db2f368a1c441e25c