Replace DataSource.run() callers
Now, `DS.check(RawData)` and `DS.run()` are duplicates.  But
`DS.check(RawData)` is more consistent with the new API as it
clarifies the distinction between fetching and checking.  We
therefore replace the calls to `run()` with calls to `fetch() ->
RawData` and `check(RawData)`.

In a following patch, we can finally remove `run()` and its callees.

CMK-5243

Change-Id: I5713c1161ad39c3645a7f62caf5b29d69f70e6cf