Drop the old ruleset overview page
The new concept has

a) The ruleset list views for different topics
b) A rule search form

Except the search there is no central view on all the rules.

Also adding the "Search rules" main menu entry with this change which
currently lists on the ruleset list page of all rulesets.

The new rule search mode will be realized with one of the next commits.

Change-Id: Ibf5d3d7e738ad62cec8d149c80f93fd2923070a0