Better separation of rule search from rule listing
Previously the two modes were mixed up, making it's relation hard to
understand. Now these are two independent pages in our hierarchy.

When opening "Search rules" it stays in the search mode and only brings
the user to the regular ruleset hierarchy when klicking on a found
ruleset title.

When opening the "Rulesets" mode, it must always be opened with a
ruleset group. The rules to be shown can be filtered using the filter
form that is available on that page, but it stays in this mode. Even if
the form looks same as the "search rules" form, the user does not leave
the "Show rules of a group" mode.

Change-Id: I9ba92c3bcc1cbdeb24d89792be6d11a7db6b87db