host_sections: remove unused function
All plugins with either node_info set to true or
extra_sections defined are migrated manually, and thus
will never be called in the legacy mode from hell.
In the remaining cases neither of the two functions
will do anything, so we remove them.

This further reduces the entaglement between host_sections
and base/config.

Change-Id: I8c53eb12f0fd1672bdf4338b718077a26fbf3fe8