11259 FIX Fix host "Save & Test" action showing "API error" for all tests
The host diagnostic page was initially showing the message "API Error: The text
None is not a valid integer number." after adding or editing a host and
confirming the dialog with the "Save & Test" button.

This regression was introduced by werk #11254 in version 1.6.0p15.

You could workaround this issue by clicking the "Test" button on the page
showing that error. It will start the tests in the correct way.

Change-Id: I8042b0bd2de35bff2d55b4cff41131147fd6a47c