Move Summarizer construction to Configurator
The checking steps are:

 # fetch() -> RawData
 # check(RawData) -> HostSections
 # summarize(HostSections) -> ServiceCheckResult

We have already moved the construction of the fetcher and of the checker
to the configurator.  It is but consistent to also move the summarizer
there.

In a next step, we might as well remove the summary out of the
Checker/Parser to better separate the three steps and clarify the
intent of the classes.

In this patch, we further notice that we really only have two different
Checker/Parsers: Agent and SNMP.  We will simplify the Checker hierarchy
in a coming patch as well.

CMK-5275

Change-Id: Ib862d63f0f36bbf83e68dfa373112fac7bca4dbe