11372 FIX mk_oracle.ps1: Fixed failed connects for 18c and newer
Oracle tends to have arbitrary output syntax for the command <tt>sqlplus
-v</tt>. This prevented the connection to databases that have 18c or newer,
because since then there are two version infos in the output string. This
will be covered with this fix. We rely on the assumption, that the first
three strings (separated by whitespace) have always the same syntax. E.g.:
SQL*Plus: Release 12.1.0.2.0.  This gives us the possibility to extract the
version from each database regardless of it's sqlplus-version.

If you have problem with your 18c-databases, please update your mk_oracle.ps1.

Change-Id: If54a8176429c55f1c2cb4798d8de8533f0b852a1