Extract loop for make_host_sections
Exception handling and host sections (summary) handling are done using
an implicit protocol as both are cached on the Checker class.

Originally, the data source class was also caching the exception from
the fetcher.  After the split, this is not practicable anymore as the
fetcher is not in the checker anymore.

Therefore, we now need to make the error handling protocol explicit.
This requires some refactoring in the data source users such as
inventory, discovery, or check_mk (compare 🍝).

The biggest culprit here is now `make_host_sections` because it calls
`fetcher.fetch()` and `checker.check()` directly and hides raw data,
exception, and host sections.

Unfortunately, the fact that `make_host_sections` takes an iterable
makes it nearly impossible to simply rearrange the code.

This patch separates the construction of the loop from the iteration on
the loop.

CMK-5275

Change-Id: I3139233dc67b03974f4bb99a7100d295ec02b1a8