Fix monitoring of clustered SNMP services
When monitoring Checkmk clusters, the cluster nodes are responsible for
fetching all information from the monitored host and cache the result
for the cluster checks to be performed on the cached information.

This means that in case of SNMP nodes, they need to take the clustered
services of the node into account, fetch the needed sections and cache
them for the cluster host.

But later, when checking the node services, the node has to only deal
with the unclustered services.

The calculation of the services is now again split into two phases (as
it was before). We first gather the services to fetch (incl. clustered)
and then the services to check (excl. clustered).

Change-Id: I7a3eb19c653955cc752c71175116ee14c28c42d9