Logwatch acknowledge: Use page menu of actual pages
Previously the "acknowledge log files" confirm page was rendered
with a completely own page header (incl. breadcrumb, page menu).
Since it is only a confirm dialog in a context of another page,
it should really have the same breadcrumb and menu.

This commit moves the confirm handling to the individual pages
which needs a bit of copy-n-paste, but that seems to be OK for
this.

Change-Id: Idbc7ad201f9eb4a66b169a8e6a402316f5e8e59e