Let mhs computation return the computed host sections
Exception handling and host sections (summary) handling are done using
an implicit protocol as both are cached on the Checker class.

Originally, the data source class was also caching the exception from
the fetcher.  After the split, this is not practicable anymore as the
fetcher is not in the checker anymore.

Therefore, we now need to make the error handling protocol explicit.
This requires some refactoring in the data source users such as
inventory, discovery, or check_mk (compare 🍝).

We modify `make_host_sections` so that it *returns* the host sections
to be summarized.

CMK-5275

Change-Id: Ie94fead164fd2ef3fea5ffc2374495ca7c838e19