Fix delayed loading of graphs
Since we introduced simplebar for the content pages, listening on scroll
elements does not work as before. We first need to finish initializing
simplebar and can then register the event listener on the scroll element
of simplebar.

To make this possible, the event listeners for the delayed loading have
to be moved to the final javascript.

Change-Id: If9877f8edf74e24082c5984c654e6ad823f41f28