Integration tests: Invalidate test image on base image changes
Once a new base image is built by our "Build Build Container" job, we
don't want to use the test containers that were built on top of the old
build containers.

This commit ensures the base image and test images are always up-to-date
or rebuilt once a new base image occurs.

It is still supported to continue with the available images in case
there is no connection possible to the docker registry.

Change-Id: Id0e2005deac506d1c321ed42bf3c1bea43c3ec10