Make summary protocol explicit in two more places
We must make the protocol for summary computation explicit in order to
be able to properly handle errors in the check helpers.

This patch demonstrates that it is possible in two more places.  The
only remaining one is in `cmk/base/inventory.py` and requires more
changes to the code.

CMK-5275

Change-Id: Ie1b384a989333b0dcd44fc386724c3baa7b940cf