Let the functions in inventory return the host_sections
This is consistent with the transformation of `make_host_sections` in
`update_host_sections` from `cmk/base/data_sources/_data_sources.py`.

See the corresponding commit for a detailed explanation.

This should be the last refactoring commit necessary to make the
implicit protocol explicit in order to fix the error handling.

CMK-5275

Change-Id: I25cf3cac23fddff7109bd9499431f1dd3f2153a2