Add handling of disabled packages to GUI
The MKP manager of the CEE can now be used to enable and disable
packages. This can be done useful when one wants to disable a package
temporarily for testing purposes. It can also be used to re-enable
packages that were disabled during update and you want to reactivate it.

Change-Id: Ib06edf678e0245374258efc19314a707bc9624ab