Minimal approach to make checker work
This change uses the exact same logic as the Check helpers. Except
that it uses the raw data from the fetcher instead of fetching new
information on it's own.

In one of the next steps we should refactor the functions changed here
(meaning check_inventory and do_check including the decorator and
update_host_sections). It is probably a good way to split them up
into separate steps to make them reusable in a cleaner way.

With enabled checkers / fetchers the Check_MK service now produces
a crash as result. This happens in some parse function because the
data transported from fetcher to checker breaks the encoding. This
will be fixed in one of the next commits.

Add some useful debug logging along the way.

Change-Id: Ic65daa545be40f00afb9e60c273f0d6b20deb3ba