Update Checker API for RTC
RTC is not in the same module as the other data sources so that it
has been forgotten during the API changes from DataSource.run() to
fetcher.fetch() / checker.check(), etc.

The class is noop anyway but we still apply the usual transformations.

This further requires generalizing the no-op class from piggyback so
that we can use it in the no-op fetcher here as well.

CMK-5275

Change-Id: Id61c54aded6ef2845b4623dc8a0cb63adf4b0fbc