Consider 1.6 feature pack packages to be outdated with 1.7
Checkmk 1.6 shipped the first feature pack MKPs which sadly had no
"version.usable_until" attribute set. To be able to disable them
automatically we use a hard coded list of package names below. All of
these packages start with the version number "1.". To ensure the known
and possible future packages are removed, we consider the known packages
to be outdated.

Change-Id: I017aa40ecee0c38b7fd283a37b7db7e5b2d1f50b