Fix service discovery display toggles
We previously tried to render the page menu once and then update single
elements of the page menu during the subsequent page updates. This did
not work well and was error prone.

We now re-render the whole page menu, just like we do it with the page
content. This consistent handling of both page elements centralizes the
logic in the Python code which makes it easier to understand and
maintain.

This commit adds the new behaviour. Following commits will simplify
the toggle elements and remove the old update logic.

Change-Id: I1dc789c9e22b36c0cac59e908b9b97bb808416c0