11365 FIX vms_if: Fix broken discovery and check execution
A bug in the check <tt>vms_if</tt>, which monitors network interfaces
on OpenVMS, prevented the corresponding services from being discovered
and resulted in a crash for already discovered services (error message:
"TypeError: '<' not supported between instances of 'str' and 'int'").
This bug occurred since 2012 and affects all hosts. With this werk,
<tt>vms_if</tt> will again be discovered and work.

This change also migrates vms_if to the new API.

CMK-5321

Change-Id: I610ab9aa6f1e20c616feef5ce1f245d2d781fa8f