Extend and reuse fetcher I/O protocol
The core already makes some minimal parsing so that errors may be
reported early.  This, however, introduces some minimal coupling
between the Python part of the protocol and its C++ part.

Therefore, it is not trivial to introduce a new protocol.

In principle, it is also in our best interest to minimalize the
number of protocols.

This patch modifies the existing header by renaming some fields and
extending others.  The changes in Python are mirrored in C++.

CMK-5316

Change-Id: I5e72e4d408b257e734f25ac7664bd92c384f777e