Make fetcher know about different execution modes
When executing a Check_MK service check, the raw data needs to be
fetched from the monitored hosts.  In case of the Check_MK Discovery
service, existing caches may be used.

Since the fetcher process cares about writing and reading of the cache,
the fetcher helper process needs to know for which if the services above
it is working.

Means: The fetcher processes work on requests of two different modes:

- Discovery mode
- Checker mode

This commit transports the mode, from the core object (Check_MK or
Check_MK Discovery service) enterprise/core/src/FetcherHelper.cc:52
to the fetcher execution cmk/fetchers/controller.py:288.

It currently does not change the actual cache handling of the fetcher.
This will be done in a later commit.

Change-Id: I187201fbf673c7151eb3aed0d71347cf2a4cb90e