10128 FIX Fixed potential monitoring core crash when rrdcached is down
The rrdtool library used by the Nagios Event Broker and the Micro Core has a
subtle bug, which could lead to undefined behavior when the rrdcached is
down or replies with an error.  Symptoms are funny log messages or
monitoring core crashes when graphs are viewed in the GUI.  This has been
fixed.

Change-Id: Ic5cf0f8452ee1f01a71caf642603eff5d5144124