11114 Add Rule set to disable individual SNMP sections
Users can now entirely disable individual SNMP sections using the rule
"Exclude SNMP sections".

As a result, the related data will not be fetched from the corresponding devices.
This may be useful if you want reduce the traffic on your network, or
suppress device responses that are known to be wrong.

This is similar to disabling check plugins, but not quite the same:
An SNMP section may be used by multiple check plugins, and it may or may
not be *required* by an individual check plugin.
Check plugins that require a section which has been disabled will not be
discovered subsequently.

CMK-5372

Change-Id: I0d0df505754dddfc0194de7f212c61834aedaf18