11343 ServiceNow: Make rule more flexible
The enterprise notification plugin "ServiceNow" has a new rule option for setting the incident state in case of an incoming recovery.
Furthermore all rules for changing the incident state can be configured by using an Integer as state. This can be useful in case the default mapping was changed by the user in ServiceNow and does not fit anymore the mapping definied in checkmk.

Change-Id: I4ab55267dc403245647d0cc3ed53650ca5f0b7ca