Hand over mode to fetcher
The cache needs to have a mode where it writes it's information to the
cache, but does not read from the cache (except in simulation mode).  In
1.6 this was realized in `DataSource._read_cache_file` by the
condition:

```
if not self._may_use_cache_file and not config.simulation_mode:
```

The flag self._may_use_cache_file was removed from the cache during some
refactorings, but in fact we still need to be able to make this decision.

This commit makes it possible again by handing over the mode known to
the caller to the `fetch` method. It then decides how to deal with the
cache.

With this change counter based checks will start working, because the
check results are not based on cached data anymore.

Change-Id: Id2587d5c4ad71d2e609283cf8ee6f5151a6e648b