Extract packed config loading from PackedConfig to existing load function
Loading the packed config does not need the PackedConfig class. We just
need to use the PackedConfigStore to read the current packed config and
merge the read definitions into the cmk.base.config. It is not necessary
to involve the PackedConfig class.

The left PackedConfig class is now a generator of the packed config
content. The next commit will clean up naming and the structure to
clarify this.

Change-Id: I3abe41fa9d13b2e4b1f46640a8a48995d461c3ef