Write packed config link to latest serial
There are some "load packed config" calls that are not aware of the
current config serial. In the past these call sites always loaded the
latest version of the packed config.

We preserve this behaviour by generating a symlink "latest" that always
points to the last packed config version. This makes the unversioned
load config calls work as before until we migrate them to be versioned.

Change-Id: Ia89b06a00921d2a04bfb19958d636f17b5d0bae5