update_rrd_fs_names: be more robust against autochecks rewriting failure
If a variable that is referenced in the autochecks file is missing, the
rewriting of the autochecks during cmk-update-config will fail.

Subsequently, the attemt to read the autochecks using the config_caches
autochecks manager will fail with a lot of missleading error messages, as the
autochecks manager assumes all references have been resolved, and does
not pass *any* variables.

Using the parse_autochecks_file function and passing the present
variables will not resolve the original issue of the missing variable,
but it will result in an exception showing the *actual* problem.

Change-Id: I11af5bf6c6f313414142cf8a1b378de1869954e4