Check-API: expect params if and only if default params are not None
We must allow check functions to have a 'params' argument even if
no ruleset is configured, as we may have discovered parameters.

To be consistent accross all plugin types we expect the functions to
accept a 'params' argument if and only if default parameters are
provided which are not None.

Change-Id: Ibfdb2c7b3d5c6d9a718c3b3575e19817a488a803