11392 FIX heartbeat_crm: strict activation in agent
After werk #11042 pacemaker-controld is a target process required to
trigger the heartbeat_crm section in Linux and openwrt agents. The way such
process was found was susceptible to name collisions leading to the
discovery of fictuos services, because other processes on the same host
could be accessing the same service.

This werk makes it more strict to catalog only pacemaker-controld processes.

Problem being the "-f" option queries after all the parameters processes
are called. Other external software or other parallel running agent could
contain pacemaker-controld as argument and lead to activation of this agent
section.

SUP-4053

Change-Id: I97c8572ca3c38851512889f6dc0694662d4c92cd