Consolidate monitoring config creation
We now have two scenarios left:

a) cmk -U: Create the core config
b) cmk -O / cmk -R: Create the core config + reload or restart of the core

Both call paths are now using the `do_create_config` function for
"create the core config" step and both are now also using the same
logic for backing up and restoring the objects file which was missing
with "cmk -U" before.

Another remaining difference is that cmk -U does not respect the
activation lock. This will be streamlined in the next commit.

Change-Id: I16a7b2394ac6375c06ac570436bfd99e0f42dff2