Rename FetcherResult -> FetcherMessage
The name `Result` is already heavily overloaded and it should take a new
meaning with the error handling.

Currently, the FetcherResult/FetcherMessage type contains a piece of
FetcherHeader and the fetcher payload in an untyped dict.  To proper
handle errors, we need the full FetcherHeader.

It is not really good practice anyway to tear objects appart.  This will
be fixed in a comming commit.

This is an automated, mechanical change using `ack | xargs sed`

CMK-5419

Change-Id: I594c5be777cc45f702e613bb6241732c1df67f24