Check plugin loading fixtures not loading automatically anymore
These fixtures must not be used automatically!  They significantly slow
down the initialization of all our unit tests, especially when only
executing a small subset of the tests. We have to be more specific.

This functionality was moved to the top-level conftest.py in 195238faf7d
with the intention to prevent plugins from being loaded multiple times.
The scope of this fixture is preventing that. This is fine.

But the autouse is not OK.

Since we have now removed "automatic loading of all checks", there were
some additional cleanups needed to make the tests work again.

Change-Id: Icd03cc0160d36e9204d88e378f0804a44ba290fe