Error handling: check & summarize
We wrap the argument to `check()` (aka `parse()` + `summarize()`) into
our `Result` type in order to encapsulate the error handling at the
appropriate level.

`fetch()` is missing as Result there needs to be serialized.

Furthermore, starting with `check()` lets us fix the API and remove
parts of the unnecessary complexity in the next commits.

CMK-5419

Change-Id: Iceca2a231b9cfc070083ce084ebb5b991e1bca1b