Improve pytest initialization time
The import of MockLiveStatusConnection makes the conftest.py load nearly the whole
GUI code. This takes some initialization overhead when running just a specific
subset of tests. We want to have a fast initialization. Quick fix: local import.

In general it is also a good idea to keep the number of dependencies low in the
conftest files.

Change-Id: I43dc63adb0c6deb600305c726e2504a24565d60e