snmp fetchers: handle SNMP detetion
The SNMP detection is now handled by the SNMP fetchers, as expected.
Hosts are no longer contacted upon config generation (CMK-5402).

Note that the persisted sections and the 'ignore_check_interval'
flag are still broken, but instead of being evaluated at the
wrong time, they are (for now) ignored completely. This issue is
tracked in CMK-5457.

Change-Id: Iabddf9581cb347956e8183fddf33bdfeb3f8fec0