11491 FIX Nagios: Fix broken config reason displaying
This issue is only affecting Checkmk installations that use the Nagios core.
In case a configuration error occurs, the reason for this issue should be displayed
in the console or the GUI like this: "When Validating Nagios configuration...ERROR: REASON".

However, since release v1.6.0p4 the output of the config validation was not shown to the user.
Instead this text was shown instead: "Validating Nagios configuration...ERROR: Configuration Error: local variable 'cmk' referenced before assignment"

Change-Id: I3e870684d1c7ce7c6e45f4b4ee370a436d02b9f2