Delegation API in Configurator
We now let the configurator (that we should now rename) handle
all the data source tasks via delegation.  This way, we obtain a
relatively simple API where the user must only know and use the
Configurator classes.

Actually, we are nearly back to the original API but with more
some renaming left to do and a more flexible design that

 > prefer[s] composition over i̵n̵h̵e̵r̵i̵t̵a̵n̵c̵e̵ god class and 🍝

CMK-5450

Change-Id: I52ee354b1929790e723e99870e7d5871dc37bfcd