11270 FIX oracle_instance: Fix missing uptime column if status data inventory is enabled
If {{Status data inventory}} is enabled via the rule {{Do hardware/software inventory}}
then {{HW/SW inventory data}} and {{status data}} are inventorized and merged in the
view {{Inventory of HOST}}. To make this work properly we have to identify the
rows in both trees. Previously we only used the {{SID}} column. In newer
{{Oracle}} agent outputs there may be several rows with the same {{SID}} but
different {{process names}}. Thus we add the {{process name}} column and use
it as an additional key for the identifier such that both trees can display the full
information of each output line.

Change-Id: I72cbfd20f2cc96279248e9b1f95519afc95dd6ea