11053 FIX Don't escape the plugin output of email notifications if configured
If you have rules configured for the rulesets "Escape HTML
codes in plugin output" or "Escape HTML codes in service
output", HTML email notifications will now honor these
settings as well.

This is e.g. usefull if you set "Enable per-process details
in long-output" to "HTML output" in a "State and count of
processes" rule. The HTML table of processes will now be
rendered correctly in email notifications as well.

Change-Id: I6c9d881f8cbf564580aa845e119fa4e54fa4067e