Transport config serial through call stack
The config serial is now gathered once in `_create_core_config` and
propagated down to all phases of the helper config creation where it is
needed.

We can now clearly see which parts of the helper config don't care about
the serial at the moment. This will be extended in a subsequent commit.

Change-Id: I3b55dc1e1636ccb0b91bbe28ac7a6218500d2d63