RRD fetch always returns TimeSeries
TimeSeries is aware of the of the time window data composes. There is no
need to carry the information of start, end, step around. The possibility
of empty graphs is handled directly on call sites.

The rest of the commit carries the consequences of always utilizing a
TimeSeries instance.

CMK-5233

Change-Id: If565770b03da8cd2a78911399c43eb032abfb8b0