Result type: mimic OCaml instead of Rust
The interface from Rust is such that

 - `Result.ok -> Optional[T_co]` and
 - `Result.err -> Optional[T_err]`

where the accessors using `Optional[]` force us to litter users with
`assert .. is not None`, even after an explicit call to `Result.is_ok()`
or `Result.is_err()`.  The alternative is to call the `unwrap()` methods
but here, the name of the function is unclear in Python.  Indeed,
`unwrap()` transfers ownership in Rust but this concept is meaningless
in Python---CPython is reference counted.

This was also noticed in code review of code using the type.

As written in the docstring, the concept of the `Result` type from Rust
exists in a variety of programming languages.  We know of three of them,
there are probably more:

 - C++ has a proposal for this and calls it `std::expected`.
 - It is the `Either` monad of Haskell (as noticed by SP).
 - OCaml calls the type `Result` as well.

The OCaml interface is the closest to the Rust interface that we found
but does not have the drawbacks identified in Rust and it is therefore
the one we select.  We do not aim to implement `join`, `bind`, `fold` or
`map` as these are more typical of a functional programming language.

This patch therefore removes the `unwrap()` methods and adds `value()`
and the rich comparison operators as defined in OCaml.

CMK-5450

Change-Id: Ie1ffe4de31826c48172183c440b493b274ea3682