Add test to for delayed host check compilation
In this mode the compiled file path links to the source file until it is
executed for the first time. Then the symlink is replaced with the
py-compiled file. Which then optimizes future executions.

The non-delayed mode compiles all files during config creation which
results in a longer config creation time.

Change-Id: I6a5093eed30313bdc8ed0b3eb56caebf1aa2c63e