API: Serialize and deserialize Fetcher in Fetcher
This consolidates the API in that the serialization and deserialization
methods are in the same class.  This also lets us separate concerns
(serialization and instantiation of the Fetcher in a Source) better.

Instantiating a Fetcher directly from a source does not perform the
useless serialization / deserialization steps anymore.

The API is also now more consistent and therefore easier to learn.

CMK-5450

Change-Id: I8d02fe363cb49a4ae3f3ce0df2fc6fdff7d4c813